Update and Data Model Schema 1.3.6
Add "isForeignNational" property to the "grantors" definition.
Add the following top-level properties: "isCheckData", "customWorkItemName", "isBuildingLoanStandalone"
Add "isBuildingLoanIntended" property to the "features" object for use with the New York Building Loan feature.
Allow "assignmentOfPermitsProperties" to be sent both for "construction" and changed the version in "features" to "cannabisAssignmentOfPermitsProperties". When the same properties is entered for both the application automatically combines both assignments into one.
Add properties to allow attribution of a MERS number for loan assignee in the "loanSale" object.
Add various properties associated with client specific customizations.
Changed various titles and descriptions for clarity and consistency.